backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.2.0
bcrypt==4.0.1
python-dotenv==1.0.0
google-cloud-firestore==2.11.1
google-auth==2.23.4
google-api-core==2.11.1